Move chunk of Minischedule.Load into Activity
